+++ CHANGELOG.md
@@ -46,6 +46,7 @@ CHANGELOG
 ### Added
 - Automatic Persisted Queries (APQ) now cache the parsed query [\#740 / mfn](https://github.com/rebing/graphql-laravel/pull/740)\
   This avoids having to re-parse the same queries over and over again.
+- Add ability to detect unused GraphQL variables [\#660 / mfn](https://github.com/rebing/graphql-laravel/pull/660)
 
 2021-04-10, 7.2.0
 -----------------